+++ app/careers/openings/[id]/page.tsx
@@ -1,334 +1,13 @@
-"use client"
-
-import type React from "react"
-
-import { useParams, notFound } from "next/navigation"
-import { useState } from "react"
-import Link from "next/link"
 import { jobOpeningsData } from "@/data/careers/openings"
-import { Button } from "@/components/ui/button"
-import { Badge } from "@/components/ui/badge"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Textarea } from "@/components/ui/textarea"
-import { ArrowLeft, MapPin, Briefcase, Clock, CheckCircle, ChevronRight, Upload } from "lucide-react"
-import { toast } from "@/components/ui/use-toast"
+import JobDetailPageClient from "./JobDetailPageClient"
+
+// This function tells Next.js which paths to pre-generate at build time
+export function generateStaticParams() {
+  return jobOpeningsData.jobs.map((job) => ({
+    id: job.id,
+  }))
+}
 
 export default function JobDetailPage() {
-  const params = useParams()
-  const jobId = params.id as string
-
-  const job = jobOpeningsData.jobs.find((job) => job.id === jobId)
-
-  const [isSubmitting, setIsSubmitting] = useState(false)
-  const [formData, setFormData] = useState({
-    fullName: "",
-    email: "",
-    phone: "",
-    coverLetter: "",
-    resume: null as File | null,
-  })
-
-  if (!job) {
-    notFound()
-  }
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    const { name, value } = e.target
-    setFormData((prev) => ({ ...prev, [name]: value }))
-  }
-
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    if (e.target.files && e.target.files[0]) {
-      setFormData((prev) => ({ ...prev, resume: e.target.files![0] }))
-    }
-  }
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault()
-    setIsSubmitting(true)
-
-    // Simulate form submission
-    setTimeout(() => {
-      setIsSubmitting(false)
-      toast({
-        title: "Application Submitted",
-        description: "Thank you for your application. We will be in touch soon.",
-      })
-
-      // Reset form
-      setFormData({
-        fullName: "",
-        email: "",
-        phone: "",
-        coverLetter: "",
-        resume: null,
-      })
-
-      // Reset file input
-      const fileInput = document.getElementById("resume") as HTMLInputElement
-      if (fileInput) fileInput.value = ""
-    }, 1500)
-  }
-
-  return (
-    <>
-      <section className="pt-32 pb-16 bg-[url('/contact-us.jpg')] bg-cover bg-center text-white relative">
-        <div className="absolute inset-0 bg-black/70 z-0"></div>
-        <div className="container mx-auto relative z-10 px-4">
-          <div className="max-w-4xl">
-            <div className="flex items-center gap-2 mb-4">
-              <Link
-                href="/careers/openings"
-                className="text-white/80 hover:text-white flex items-center gap-1 transition-colors"
-              >
-                <ArrowLeft className="h-4 w-4" />
-                Back to Job Openings
-              </Link>
-            </div>
-            <h1 className="text-4xl md:text-5xl font-bold mb-4">{job.title}</h1>
-            <div className="flex flex-wrap gap-y-2 gap-x-4 mb-4 text-white/80">
-              <div className="flex items-center">
-                <Briefcase className="h-4 w-4 mr-2" />
-                {job.department}
-              </div>
-              <div className="flex items-center">
-                <MapPin className="h-4 w-4 mr-2" />
-                {job.location}
-              </div>
-              <div className="flex items-center">
-                <Clock className="h-4 w-4 mr-2" />
-                Posted {job.postedDate}
-              </div>
-            </div>
-            <Badge variant={job.type === "Full-time" ? "default" : "secondary"} className="mt-2">
-              {job.type}
-            </Badge>
-          </div>
-        </div>
-      </section>
-
-      <section className="py-16">
-        <div className="container mx-auto px-4">
-          <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
-            <div className="lg:col-span-2">
-              <Card>
-                <CardHeader>
-                  <CardTitle>Job Description</CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <p className="text-muted-foreground mb-8">{job.description}</p>
-
-                  <h3 className="text-xl font-bold mb-4">Requirements</h3>
-                  <ul className="space-y-3 mb-8">
-                    {job.requirements.map((requirement, index) => (
-                      <li key={index} className="flex items-start">
-                        <CheckCircle className="h-5 w-5 text-primary mr-3 flex-shrink-0 mt-0.5" />
-                        <span>{requirement}</span>
-                      </li>
-                    ))}
-                  </ul>
-
-                  <h3 className="text-xl font-bold mb-4">Our Recruitment Process</h3>
-                  <ol className="space-y-3 mb-8">
-                    {jobOpeningsData.recruitmentProcess.map((step, index) => (
-                      <li key={index} className="flex items-start">
-                        <span className="bg-primary text-white rounded-full w-6 h-6 flex items-center justify-center font-bold mr-3 flex-shrink-0">
-                          {index + 1}
-                        </span>
-                        <span>{step}</span>
-                      </li>
-                    ))}
-                  </ol>
-                </CardContent>
-              </Card>
-            </div>
-
-            <div>
-              <Card>
-                <CardHeader>
-                  <CardTitle>Apply for this Position</CardTitle>
-                  <CardDescription>Fill out the form below to apply for the {job.title} position</CardDescription>
-                </CardHeader>
-                <CardContent>
-                  <form onSubmit={handleSubmit}>
-                    <div className="grid gap-4">
-                      <div className="grid gap-2">
-                        <Label htmlFor="fullName">Full Name</Label>
-                        <Input
-                          id="fullName"
-                          name="fullName"
-                          placeholder="John Doe"
-                          required
-                          value={formData.fullName}
-                          onChange={handleInputChange}
-                        />
-                      </div>
-
-                      <div className="grid gap-2">
-                        <Label htmlFor="email">Email</Label>
-                        <Input
-                          id="email"
-                          name="email"
-                          type="email"
-                          placeholder="john.doe@example.com"
-                          required
-                          value={formData.email}
-                          onChange={handleInputChange}
-                        />
-                      </div>
-
-                      <div className="grid gap-2">
-                        <Label htmlFor="phone">Phone Number</Label>
-                        <Input
-                          id="phone"
-                          name="phone"
-                          placeholder="+1 (555) 123-4567"
-                          required
-                          value={formData.phone}
-                          onChange={handleInputChange}
-                        />
-                      </div>
-
-                      <div className="grid gap-2">
-                        <Label htmlFor="resume">Resume/CV</Label>
-                        <div className="border border-input rounded-md p-2">
-                          <Label
-                            htmlFor="resume"
-                            className="flex items-center gap-2 cursor-pointer hover:text-primary transition-colors"
-                          >
-                            <Upload className="h-5 w-5" />
-                            {formData.resume ? formData.resume.name : "Upload your resume (PDF, DOC, DOCX)"}
-                          </Label>
-                          <Input
-                            id="resume"
-                            name="resume"
-                            type="file"
-                            accept=".pdf,.doc,.docx"
-                            required
-                            className="hidden"
-                            onChange={handleFileChange}
-                          />
-                        </div>
-                      </div>
-
-                      <div className="grid gap-2">
-                        <Label htmlFor="coverLetter">Cover Letter</Label>
-                        <Textarea
-                          id="coverLetter"
-                          name="coverLetter"
-                          placeholder="Tell us why you're interested in this position and what makes you a great fit..."
-                          rows={5}
-                          value={formData.coverLetter}
-                          onChange={handleInputChange}
-                        />
-                      </div>
-                    </div>
-
-                    <Button type="submit" className="w-full mt-6" disabled={isSubmitting}>
-                      {isSubmitting ? "Submitting..." : "Submit Application"}
-                    </Button>
-                  </form>
-                </CardContent>
-              </Card>
-
-              <Card className="mt-6">
-                <CardHeader>
-                  <CardTitle>Share This Job</CardTitle>
-                </CardHeader>
-                <CardContent>
-                  <div className="flex gap-2">
-                    <Button variant="outline" size="icon" className="rounded-full">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="16"
-                        height="16"
-                        fill="currentColor"
-                        viewBox="0 0 16 16"
-                      >
-                        <path d="M16 8.049c0-4.446-3.582-8.05-8-8.05C3.58 0-.002 3.603-.002 8.05c0 4.017 2.926 7.347 6.75 7.951v-5.625h-2.03V8.05H6.75V6.275c0-2.017 1.195-3.131 3.022-3.131.876 0 1.791.157 1.791.157v1.98h-1.009c-.993 0-1.303.621-1.303 1.258v1.51h2.218l-.354 2.326H9.25V16c3.824-.604 6.75-3.934 6.75-7.951z" />
-                      </svg>
-                    </Button>
-                    <Button variant="outline" size="icon" className="rounded-full">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="16"
-                        height="16"
-                        fill="currentColor"
-                        viewBox="0 0 16 16"
-                      >
-                        <path d="M5.026 15c6.038 0 9.341-5.003 9.341-9.334 0-.14 0-.282-.006-.422A6.685 6.685 0 0 0 16 3.542a6.658 6.658 0 0 1-1.889.518 3.301 3.301 0 0 0 1.447-1.817 6.533 6.533 0 0 1-2.087.793A3.286 3.286 0 0 0 7.875 6.03a9.325 9.325 0 0 1-6.767-3.429 3.289 3.289 0 0 0 1.018 4.382A3.323 3.323 0 0 1 .64 6.575v.045a3.288 3.288 0 0 0 2.632 3.218 3.203 3.203 0 0 1-.865.115 3.23 3.23 0 0 1-.614-.057 3.283 3.283 0 0 0 3.067 2.277A6.588 6.588 0 0 1 .78 13.58a6.32 6.32 0 0 1-.78-.045A9.344 9.344 0 0 0 5.026 15z" />
-                      </svg>
-                    </Button>
-                    <Button variant="outline" size="icon" className="rounded-full">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="16"
-                        height="16"
-                        fill="currentColor"
-                        viewBox="0 0 16 16"
-                      >
-                        <path d="M0 1.146C0 .513.526 0 1.175 0h13.65C15.474 0 16 .513 16 1.146v13.708c0 .633-.526 1.146-1.175 1.146H1.175C.526 16 0 15.487 0 14.854V1.146zm4.943 12.248V6.169H2.542v7.225h2.401zm-1.2-8.212c.837 0 1.358-.554 1.358-1.248-.015-.709-.52-1.248-1.342-1.248-.822 0-1.359.54-1.359 1.248 0 .694.521 1.248 1.327 1.248h.016zm4.908 8.212V9.359c0-.216.016-.432.08-.586.173-.431.568-.878 1.232-.878.869 0 1.216.662 1.216 1.634v3.865h2.401V9.25c0-2.22-1.184-3.252-2.764-3.252-1.274 0-1.845.7-2.165 1.193v.025h-.016a5.54 5.54 0 0 1 .016-.025V6.169h-2.4c.03.678 0 7.225 0 7.225h2.4z" />
-                      </svg>
-                    </Button>
-                    <Button variant="outline" size="icon" className="rounded-full">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="16"
-                        height="16"
-                        fill="currentColor"
-                        viewBox="0 0 16 16"
-                      >
-                        <path d="M13.601 2.326A7.854 7.854 0 0 0 7.994 0C3.627 0 .068 3.558.064 7.926c0 1.399.366 2.76 1.057 3.965L0 16l4.204-1.102a7.933 7.933 0 0 0 3.79.965h.004c4.368 0 7.926-3.558 7.93-7.93A7.898 7.898 0 0 0 13.6 2.326zM7.994 14.521a6.573 6.573 0 0 1-3.356-.92l-.24-.144-2.494.654.666-2.433-.156-.251a6.56 6.56 0 0 1-1.007-3.505c0-3.626 2.957-6.584 6.591-6.584a6.56 6.56 0 0 1 4.66 1.931 6.557 6.557 0 0 1 1.928 4.66c-.004 3.639-2.961 6.592-6.592 6.592zm3.615-4.934c-.197-.099-1.17-.578-1.353-.646-.182-.065-.315-.099-.445.099-.133.197-.513.646-.627.775-.114.133-.232.148-.43.05-.197-.1-.836-.308-1.592-.985-.59-.525-.985-1.175-1.103-1.372-.114-.198-.011-.304.088-.403.087-.088.197-.232.296-.346.1-.114.133-.198.198-.33.065-.134.034-.248-.015-.347-.05-.099-.445-1.076-.612-1.47-.16-.389-.323-.335-.445-.34-.114-.007-.247-.007-.38-.007a.729.729 0 0 0-.529.247c-.182.198-.691.677-.691 1.654 0 .977.71 1.916.81 2.049.098.133 1.394 2.132 3.383 2.992.47.205.84.326 1.129.418.475.152.904.129 1.246.08.38-.058 1.171-.48 1.338-.943.164-.464.164-.86.114-.943-.049-.084-.182-.133-.38-.232z" />
-                      </svg>
-                    </Button>
-                    <Button variant="outline" size="icon" className="rounded-full">
-                      <svg
-                        xmlns="http://www.w3.org/2000/svg"
-                        width="16"
-                        height="16"
-                        fill="currentColor"
-                        viewBox="0 0 16 16"
-                      >
-                        <path d="M15.545 6.558a9.42 9.42 0 0 1 .139 1.626c0 2.434-.87 4.492-2.384 5.885h.002C11.978 15.292 10.158 16 8 16A8 8 0 1 1 8 0a7.689 7.689 0 0 1 5.352 2.082l-2.284 2.284A4.347 4.347 0 0 0 8 3.166c-2.087 0-3.86 1.408-4.492 3.304a4.792 4.792 0 0 0 0 3.063h.003c.635 1.893 2.405 3.301 4.492 3.301 1.078 0 2.004-.276 2.722-.764h-.003a3.702 3.702 0 0 0 1.599-2.431H8v-3.08h7.545z" />
-                      </svg>
-                    </Button>
-                  </div>
-                </CardContent>
-              </Card>
-
-              <Card className="mt-6">
-                <CardHeader>
-                  <CardTitle>Similar Positions</CardTitle>
-                </CardHeader>
-                <CardContent className="p-0">
-                  <div className="divide-y">
-                    {jobOpeningsData.jobs
-                      .filter((j) => j.id !== job.id && j.department === job.department)
-                      .slice(0, 3)
-                      .map((similarJob, index) => (
-                        <Link
-                          key={index}
-                          href={`/careers/openings/${similarJob.id}`}
-                          className="block p-4 hover:bg-muted transition-colors"
-                        >
-                          <div className="flex justify-between items-center">
-                            <div>
-                              <h4 className="font-medium">{similarJob.title}</h4>
-                              <p className="text-sm text-muted-foreground">{similarJob.location}</p>
-                            </div>
-                            <ChevronRight className="h-5 w-5 text-muted-foreground" />
-                          </div>
-                        </Link>
-                      ))}
-                  </div>
-                </CardContent>
-              </Card>
-            </div>
-          </div>
-        </div>
-      </section>
-    </>
-  )
+  return <JobDetailPageClient />
 }
