+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "export": "next export"
   },
   "dependencies": {
     "@emotion/is-prop-valid": "latest",
@@ -41,6 +42,7 @@
     "@react-email/components": "^0.0.36",
     "@react-email/render": "^1.0.6",
     "@react-email/tailwind": "^1.0.4",
+    "@supabase/supabase-js": "^2.49.4",
     "autoprefixer": "^10.4.20",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
